refactor: Extract team search endpoint & types
