+++ packages/api-client/src/user/Contact.ts
@@ -17,15 +17,6 @@
  *
  */
 
-import type {Role} from '../team';
-import type {ManagedSource} from './ManagedSource';
 import type {User} from './User';
 
-export type Contact = Pick<User, 'id' | 'email' | 'accent_id' | 'handle' | 'name' | 'team'> & {
-  /** Timestamp of invitation creation. */
-  created_at?: string;
-  managed_by?: ManagedSource;
-  role?: Role;
-  /** URL of the SAML identity provider. */
-  saml_idp?: string;
-};
+export type Contact = Pick<User, 'accent_id' | 'handle' | 'id' | 'name' | 'team'>;
