+++ packages/api-client/src/user/UserAPI.ts
@@ -36,7 +36,6 @@ import type {
   VerifyDelete,
 } from '../user/';
 import {RequestCancellationError} from './UserError';
-import type {SearchOptions} from './SearchOptions';
 import type {ClientPreKey, PreKeyBundle, QualifiedPreKeyBundle} from '../auth/';
 import type {PublicClient} from '../client/';
 import type {RichInfo} from './RichInfo';
@@ -57,7 +56,6 @@ export class UserAPI {
     PROPERTIES: '/properties',
     RICH_INFO: 'rich-info',
     SEARCH: '/search',
-    BROWSE_TEAM: 'browse-team',
     SEND: 'send',
     USERS: '/users',
   };
@@ -236,44 +234,6 @@ export class UserAPI {
     if (limit) {
       config.params.size = limit;
     }
-
-    const handleRequest = async () => {
-      try {
-        const response = await this.client.sendJSON<SearchResult>(config);
-        return response.data;
-      } catch (error) {
-        if (error.message === SyntheticErrorLabel.REQUEST_CANCELLED) {
-          throw new RequestCancellationError('Search request got cancelled');
-        }
-        throw error;
-      }
-    };
-
-    return {
-      cancel: () => cancelSource.cancel(SyntheticErrorLabel.REQUEST_CANCELLED),
-      response: handleRequest(),
-    };
-  }
-
-
-  /**
-   * Search for team members.
-   * @param query The search query
-   * @param options Search options (sort, order, filter, etc.)
-   * @see https://staging-nginz-https.zinfra.io/swagger-ui/#!/users/search
-   */
-  public async getSearchMembers(teamId: string, query: string, options: SearchOptions = {}): Promise<RequestCancelable<SearchResult>> {
-    const cancelSource = Axios.CancelToken.source();
-    const config: AxiosRequestConfig = {
-      cancelToken: cancelSource.token,
-      method: 'get',
-      params: {
-        q: query,
-        ...options,
-        frole: options.frole?.join(','),
-      },
-      url: `/${UserAPI.URL.CONTACTS}/${UserAPI.URL.BROWSE_TEAM}/${teamId}`,
-    };
 
     const handleRequest = async () => {
       try {
